test(backend): move pagination tests into payment pointer service
